refactor(oceans-e2e): split OceansPage.ts into one file per POM
OceansPage.ts → OceansPage (base) + AppMode
TrainingPage.ts → TrainingPage (abstract)
FishVTrashPage.ts → FishVTrashPage
FishShortPage.ts → FishShortPage

Each spec imports only what it needs. All 41 tests pass.
