+++ frontend/packages/labs/oceans/e2e/OceansPage.ts
@@ -1,4 +1,4 @@
-import {type Locator, type Page, expect} from 'playwright/test';
+import {type Locator, type Page} from 'playwright/test';
 
 /** App mode values mirrored from `src/oceans/constants.ts`. */
 export const AppMode = {
@@ -209,147 +209,5 @@ export class OceansPage {
   async runPrediction(timeout = 15_000): Promise<void> {
     await this.runButton.click();
     await this.predictContinueButton.waitFor({state: 'visible', timeout});
-  }
-}
-
-/**
- * Base for modes that include the Training scene.
- *
- * Subclasses supply the mode-specific yes/no button locators; this class
- * provides `classifyOne`, `train`, and `fullFlow` built on top of them.
- */
-export abstract class TrainingPage extends OceansPage {
-  /** Positive-answer training button (label depends on mode / selected word). */
-  abstract get yesButton(): Locator;
-
-  /** Negative-answer training button (label depends on mode / selected word). */
-  abstract get noButton(): Locator;
-
-  /**
-   * Click Yes (or No) once and wait until the training counter increments.
-   *
-   * `onClassifyFish` increments the count synchronously but also sets
-   * `isRunning: true`. While `isRunning` is true (~1 s fish-eat animation)
-   * subsequent clicks are silently ignored. This method retries the click until
-   * the counter advances, so callers never need to manage animation timing.
-   *
-   * @param isYes - `true` to click Yes, `false` to click No.
-   */
-  async classifyOne(isYes: boolean): Promise<void> {
-    const currentText = (await this.trainCount.textContent()) ?? '0';
-    const count = parseInt(currentText.trim(), 10);
-    const next = Math.min(999, count + 1);
-
-    await expect(async () => {
-      if (isYes) {
-        await this.yesButton.click();
-      } else {
-        await this.noButton.click();
-      }
-      await expect(this.trainCount).toHaveText(String(next), {timeout: 1_200});
-    }).toPass({timeout: 5_000});
-  }
-
-  /**
-   * Perform a training sequence of yes-clicks followed by no-clicks.
-   *
-   * @param yesCount - Number of Yes classifications to submit.
-   * @param noCount  - Number of No classifications to submit.
-   */
-  async train(yesCount: number, noCount: number): Promise<void> {
-    for (let i = 0; i < yesCount; i++) {
-      await this.classifyOne(true);
-    }
-    for (let i = 0; i < noCount; i++) {
-      await this.classifyOne(false);
-    }
-  }
-
-  /**
-   * Complete the Training → Predicting → Pond flow.
-   *
-   * Trains `yesCount` yes + `noCount` no examples (default: 0 each), then
-   * runs prediction and advances to the Pond scene.
-   *
-   * @param yesCount - Yes classifications to submit before advancing.
-   * @param noCount  - No classifications to submit before advancing.
-   */
-  async fullFlow(yesCount = 0, noCount = 0): Promise<void> {
-    await this.train(yesCount, noCount);
-    await this.advanceToPredictScene();
-    await this.runPrediction();
-    await this.predictContinueButton.click();
-    await this.waitForPondScene();
-  }
-}
-
-/**
- * Page object for FishVTrash mode.
- *
- * Yes button label: "Fish". No button label: "Not Fish".
- */
-export class FishVTrashPage extends TrainingPage {
-  /** "Fish" button — exact match so it doesn't also resolve "Not Fish". */
-  get yesButton(): Locator {
-    return this.page.getByRole('button', {name: 'Fish', exact: true});
-  }
-
-  get noButton(): Locator {
-    return this.getButton('Not Fish');
-  }
-
-  /**
-   * Navigate to FishVTrash and wait for the training scene.
-   *
-   * @param page - Playwright Page fixture.
-   */
-  static async load(page: Page): Promise<FishVTrashPage> {
-    const p = new FishVTrashPage(page);
-    await p.goto(AppMode.FishVTrash);
-    await p.waitForTrainingScene();
-    return p;
-  }
-}
-
-/**
- * Page object for FishShort (and FishLong) mode after a word has been selected.
- *
- * Yes button label: the selected word. No button label: "Not <word>".
- */
-export class FishShortPage extends TrainingPage {
-  /**
-   * @param page - Playwright Page fixture.
-   * @param word - The word selected in the Words scene (e.g. "Blue").
-   */
-  constructor(
-    page: Page,
-    readonly word: string,
-  ) {
-    super(page);
-  }
-
-  /** Yes button — exact match so it doesn't also resolve "Not <word>". */
-  get yesButton(): Locator {
-    return this.page.getByRole('button', {name: this.word, exact: true});
-  }
-
-  get noButton(): Locator {
-    return this.getButton(`Not ${this.word}`);
-  }
-
-  /**
-   * Navigate to FishShort, wait for the Words scene, click the given word,
-   * then wait for the Training scene.
-   *
-   * @param page - Playwright Page fixture.
-   * @param word - Word label to select (e.g. "Blue").
-   */
-  static async load(page: Page, word: string): Promise<FishShortPage> {
-    const p = new FishShortPage(page, word);
-    await p.goto(AppMode.FishShort);
-    await p.waitForWordsScene();
-    await p.getButton(word).click();
-    await p.waitForTrainingScene();
-    return p;
   }
 }
