+++ frontend/packages/labs/oceans/e2e/fishvtrash.spec.ts
@@ -1,6 +1,7 @@
 import {expect, test} from 'playwright/test';
 
-import {FishVTrashPage, OceansPage} from './OceansPage';
+import {FishVTrashPage} from './FishVTrashPage';
+import {OceansPage} from './OceansPage';
 
 // ─────────────────────────────────────────────────────────────────────────────
 // FishVTrash — Training scene
